pyproject.toml
[tool.poetry]
name = "ufsc-update-notifier"
version = "0.1.0"
description = ""
authors = ["TiagoSansao <tiagossansao@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
requests = "^2.31.0"
python = "^3.10"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ufsc-update-notifier = "src.main:main"